+++ src/xtd.drawing.native.fltk/src/xtd/drawing/native/fltk/pen.cpp
@@ -3,18 +3,18 @@
 
 using namespace xtd::drawing::native;
 
-intptr_t pen::create() {
-  return reinterpret_cast<intptr_t>(new fl_pen());
+intmax_t pen::create() {
+  return reinterpret_cast<intmax_t>(new fl_pen());
 }
 
-void pen::color(intptr_t pen, uint8_t a, uint8_t r, uint8_t g, uint8_t b) {
+void pen::color(intmax_t pen, uint_least8_t a, uint_least8_t r, uint_least8_t g, uint_least8_t b) {
   reinterpret_cast<fl_pen*>(pen)->color(fl_rgb_color(r, g, b));
 }
 
-void pen::width(intptr_t pen, float width) {
+void pen::width(intmax_t pen, float width) {
   reinterpret_cast<fl_pen*>(pen)->width(static_cast<int>(width));
 }
 
-void pen::destroy(intptr_t pen) {
+void pen::destroy(intmax_t pen) {
   delete reinterpret_cast<fl_pen*>(pen);
 }
